### 2023-04-20 #102 @katoss
cardsort: analyzing data from open card sorting tasks
**Submitting Author:** (@katoss)
**All current maintainers:** (@katoss)
Package Name: cardsort
**One-Line Description of Package:** A python package to analyse data from open card sorting tasks
**Repository Link:**  [https://github.com/katoss/cardsort](https://github.com/katoss/cardsort)
**Version submitted:**  0.2.2
**Reviewer 1:**  @Robaina
**Reviewer 2:** @khynder
**Archive:** https://zenodo.org/record/8214655
**Version accepted:** 0.2.36
**Date accepted (month/day/year):** 2023/08/16

---

Card sorting is a standard research method in the human computer interaction field to design information architectures. The` cardsort` package analyzes and visualizes data from open card sorting tasks. Using csv data in the format returned by the [kardsort](https://kardsort.com/) tool (or any other tool outputting the same columns), it outputs a dendrogram based on hierarchical cluster analysis and pairwise similarity scores. It can also return category labels to learn which labels study participants gave to combinations of cards from emerging clusters.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience are researchers or practitioners in human computer interaction and user experience. Open card sorting is a [popular user research method](https://dl.acm.org/doi/abs/10.1145/3411763.3451750) used to understand how people order and conceptualize information, in order to design information architectures. In order to make sense of the data, clusters are often visualized in form of a [dendrogram](https://uxpajournal.org/card-sort-analysis-best-practices-2/). To do so, pairwise similarity scores need to be calculated for all cards, followed by a hierarchical cluster analysis. This functionality is provided by the cardsort package. It also offers functionality to return category labels, in order to learn which labels study participants gave to combinations of cards in the emerging clusters.

I did not find any python packages, nor other open source tools, that accomplish this, which was my motivation to make this package, when I was running a card sorting study in the course of my PhD. While research articles describe the steps of this analysis (i.e. pairwise similarity scores --> hierarchical cluster analysis --> dendrogram, see articles linked in question above), I did not find any open source tools that help to do this analysis. I used the widely used [kardsort.com](https://kardsort.com/) tool to collect the data, which refers to rather dated, closed source Windows software for analysis, which underpins my assumption of a lack of open source tools. The approach is rather simple, making use of scipy for hierarchical cluster analysis and dendrogram, adding a custom function to create the similarity scores, and putting everything in an easy-to-use pipeline. Nevertheless, in doing so, I think the cardsort package can help remove barriers for the application of this user research method (easy to use even for python beginners, no need to use closed source software that only runs on windows or expensive subscription tools). In any case I would have liked to have this package when I started my study :)

#101
