### 2023-04-22 #102 @Batalex
- [ ] The package imports properly into a standard Python environment `import package-name`.
  - [ ] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
  - [ ] **Code of Conduct** The package has a `Code of Conduct` file.
- [ ] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---
