### 2023-04-24 #102 @katoss
Thank you @Batalex ! I will try to address your comments in the course of the week. :)
