### 2023-04-26 19d6a5d
fix: move myst-nb to dev dependencies
