+++ pyproject.toml
@@ -12,7 +12,6 @@ pandas = "^1.5.3"
 numpy = "^1.24.2"
 matplotlib = "^3.7.1"
 scipy = "^1.10.1"
-myst-nb = "^0.17.1"
 
 [tool.poetry.dev-dependencies]
 
@@ -20,10 +19,10 @@ myst-nb = "^0.17.1"
 pytest = "^7.2.2"
 pytest-cov = "^4.0.0"
 jupyter = "^1.0.0"
-myst-nb = {version = "^0.17.1", python = "^3.9"}
 sphinx-autoapi = "^2.0.1"
 sphinx-rtd-theme = "^1.2.0"
 python-semantic-release = "^7.33.2"
+myst-nb = "^0.17.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
