### 2023-04-26 a4ffd3c
fix: fix version variable and source to allow automatic bumping by semantic release
