### 2023-04-26 b1816b8
Revert "fix: fix version variable and source to allow automatic bumping by semantic release"
This reverts commit a4ffd3cc1740aac8b40a1fb2e5a24e86d08f342d.

commit broke publishing to new package version with semantic release
