### 2023-04-27 #102 @katoss
Hi @Batalex,
thank you very much for your feedback. I tried addressing everything, but have some open questions:

> The version in pyproject.toml does not match the latest release. Is there any way to use semantic release to update the pyproject.toml file based on the tag? If not, maybe we should set the version here to something like "0.0.0" to indicate that this is not the true version.

Normally semantic release should bump the version automatically, but I keep running into issues. It only seems to work via getting the release version from the GitHub tag. I set the version to 0.0.0 for now, as you suggested. I will try to find a better solution, but for now at least the version is correctly updated on github and pypi.

> This is on the verge of reviewing the code, so I will no go any further at the moment, but the dependencies constraints are quite hard.
pandas = "^1.5.3"
numpy = "^1.24.2"
matplotlib = "^3.7.1"
scipy = "^1.10.1"
You could increase the pool of potential users by loosening the constraints. Based on what I saw in the sources, we do not need the latest versions for these four packages.

I am sorry if this is a stupid question, but how do I know which are the minimum versions of packages I can allow without breaking my code? I tried using `pipdeptree`, but did not find that very helpful, because it only tells me which are the dependencies of the currently installed package versions, not if I could use something lower.

> + myst-nb should be a dev dependency

True, thank you. Fixed it.

> I believe the import would be clearer if we just used the functionalities from the top module namespace, e.g from cardsort import get_distance_matrix or import cardsort; cardsort.get_distance_matrix() instead of propagating analysis all the way up to the user code. This is quite easy, you just need to import the functions in __init__.py

I wasn't aware that this was possible. Thank you, fixed it!

Best,

@katoss
