### 2023-04-28 1282da6
fix: lower dependency requirements for pandas, numpy, matplotlib, and scipy
