+++ pyproject.toml
@@ -8,10 +8,10 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.8,<3.12"
-pandas = "^1.5.3"
-numpy = "^1.24.2"
-matplotlib = "^3.7.1"
-scipy = "^1.10.1"
+pandas = "^1.5.0"
+numpy = "^1.20.3"
+matplotlib = "^3.3.2"
+scipy = "^1.7.1"
 
 [tool.poetry.dev-dependencies]
 
@@ -23,6 +23,7 @@ sphinx-autoapi = "^2.0.1"
 sphinx-rtd-theme = "^1.2.0"
 python-semantic-release = "^7.33.2"
 myst-nb = "^0.17.2"
+pipdeptree = "^2.7.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
