+++ pyproject.toml
@@ -30,8 +30,8 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.semantic_release]
-version_variable = "pyproject.toml:tool.poetry.version" # version location
-version_source = "tag"                      # getting release version from GitHub tag
+version_variable = "cardsort/__init__.py:__version__" # version location
+version_source = "commit"                      # getting release version from GitHub tag
 branch = "main"                             # branch to make releases of
 changelog_file = "CHANGELOG.md"             # changelog file
 build_command = "poetry build"              # build dists
