### 2023-04-28 12a3570
fix: fix version variable path
