### 2023-04-28 3e2c5c9
fix: fix version variable path
