### 2023-04-28 8f787fb
fix: try to get version variable from poetry commit
