### 2023-04-28 cb2b272
fix: trying to fix access to version variable for semantic release
