### 2023-04-28 898b951
fix: try to fix version variable for semantic release
