### 2023-04-28 d573bdb
fix: try to fix path to version
