### 2023-04-28 cde28a2
fix: use github tag to get version
