+++ pyproject.toml
@@ -23,7 +23,6 @@ sphinx-autoapi = "^2.0.1"
 sphinx-rtd-theme = "^1.2.0"
 python-semantic-release = "^7.33.2"
 myst-nb = "^0.17.2"
-pipdeptree = "^2.7.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
