### 2023-05-10 #102 @Batalex
:wave: Hi @robaina and @khynder! Thank you for volunteering to review for pyOpenSci! I am thrilled to have you both on board for this review. Feel free to introduce yourselves to our very own submitter @katoss 🤗

@katoss, meet @robaina, whom I had the pleasure of working with on his submission to pyOpenSci.
As for @khynder, I have enjoyed working with her over the past five years and can vouch for her dedication and passion for science.

Please get in touch with any questions or concerns! Your review is due: on June 5th.
---

Reviewers: @robaina, @khynder
Due date: June 5th, 2023
