### 2023-05-10 #102 @Robaina
Hi @katoss, nice to meet you! I'll be one of the reviewers of `cardsort`. I had such a nice experience when I submitted my package to pyOpenSci. Hope you do too! Looking forward to the review.
