### 2023-05-14 #102 @Robaina
Hi @Batalex, @katoss,

a quick question: according to [pyOpenSci docs](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#:~:text=Be%20sure%20that,you%20should%20install!) I understand that I should review `cardsort v0.2.2` and not the latest release (v0.2.12) since this is the version that was submitted to pyOpenSci, right? Just to be sure.
