### 2023-05-15 #102 @Batalex
Hey,
This is indeed a tricky question. On the one hand, it would be inconvenient to ask submitting authors to freeze the project development during submission. On the other hand, reviewing the complete code base is already very time-consuming for the reviewers; they do not have the time to review the whole code repeatedly.
We are going to proceed with the following approach:
- @Robaina, @khynder **the version listed (0.2.2) is the one you should review**
- Once you are done with your review, it is up to @katoss to point out the PRs relevant to your comments / requested changes.

The rationale behind this approach is that we want to ensure the review scope stays more or less the same. No one wants to commit to a review on a fixed scope only to discover that the goal post changed a few weeks later.
Furthermore, some of the reviewers' comments might already be addressed in the development branch. The submitting author would then be the most likely to answer with something like "thanks for the comment. Are you satisfied with the solution we made a few days ago on ecb645 / PR # 32?"

Based on what I saw, the most recent releases deal with a few mishaps on the packaging side. Most of the code base remains untouched so this approach should be fine!
