### 2023-05-17 #102 @Robaina
⭐ I have completed the review ⭐

I find `cardsort`a nice contribution! You have put a lot of work into structuring and documenting the package, it has sufficient tests and a cool CI / CD pipeline 🚀🚀 . I have got some comments, mostly suggestions to make the code more efficient and readable. I hope these will be useful to you! Please, let me know if something isn't clear from my comments. Also, I'm willing to help you implement these suggestions with PRs if you want.

Ok, here it goes:

__CHECKLIST__:

It is succinct but sufficient. However, I think the repo would benefit from a little more explanation for non-experts. This may be as easy as providing a link to an url explaining why card sorting is useful in UX design.

The instructions are located within the 'CONTRIBUTING.md' file. I would add the link to this file directly in the Contributing section of the README.md file.

The documentation is extensive. Hosted in readthedocs.io. I suggest to provided the url to the docs in the About section of the repo to increase visibility.

I suggest adding the link to the contributing file in the contributing section of the README.md file.

The current name of the package, 'cardsort', is missing as such in the title of the README file. Also, if you like, you could create a simple logo for the package to attract more users (See [this repo](https://github.com/vocalpy/crowsetta) as an example).

There are no badges in the README. You can read about badges [here](https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html#what-your-readme-md-file-should-contain) in case you are not familiar. Also, I can help with this if you need it.

As stated above, I suggest adding the URL to the About section of the repo to increase visibility.

No citation info was provided. I suggest publishing the repo in [Zenodo](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) and adding the citation and the DOI to the README.md file. Once published, Zenodo provides a badge containing the DOI and which you can display in the badge section of the README file. You could also add a [CITATON.cff](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files) file to the repo so users can easily get the citation string.

While the purpose of this package is indicated in the README, I think a little more explanation for non-experts would be useful. This may be as easy as providing a link to an url explaining why card sorting is useful in UX design.

As long as I can tell, the package works as expected.

    There are no linting tests in the CI. Also, the original code does not follow PEP8 in some cases (see below). I suggest using a linter such as [black](https://github.com/psf/black) or [ruff](https://github.com/charliermarsh/ruff) (highly recommended) to ensure the code follows PEP8.

---

__General comments__:

1. While not necessary, I think using [type hints](https://docs.python.org/3/library/typing.html) would improve the readability of the functions.

2. Inconsistent use of `.query` and `.loc` as selectors. I would recommend using `.loc` in all cases as the code is not using complex queries and `.loc` may be more readable for Python users.

3. PEP-8 recommends using format strings instead of concatenation as it improves readability. Overall, I would recommend using a linter such as [black](https://github.com/psf/black) or [ruff](https://github.com/charliermarsh/ruff) (highly recommended) to ensure the code follows PEP8. The linter could also be added to a CI pipeline (let me know if you need help with this)

4. There are several cases in which a `while` loop followed by a `try-except` block is used. This choice makes the code more complex and it is unnecessary. In these cases, it can be replaced by a simpler `loop`, removing the need for the `try-except` block (see below).

5. 'example-data.csv`is located within /docs, you could indicate the relative path in the example to help users find the file within the repo.

__Detailed comments for each function__:

_get_distance_matrix_for_user
```

1. There is no need to update the index variables, ```i, j```, within the loop, their value is set by the for loop itself.

2. Given that the distance matrix is symmetric, you could save computational time by only calculating the upper triangle of the matrix and then copying the values to the lower triangle. For example:

for i in range(n):
    for j in range(i, n):
        cat1 = ...
        cat2 = ...
        X[i, j] = X[j, i] = 0 if cat1 == cat2 else 1
```
Note that you could also write the if - else statement in a single line as shown above.

3. Perhaps using `loc` instead of `query` would be more readable for Python users, we are not making a complex query here anyway.

get_distance_matrix
```

1. You could replace the while loop with a for loop and avoid the need to update the index variable, id, within the loop. To this end, you could first find all unique user_ids in the dataframe: ```user_ids = df['user_id'].unique()```, and then loop over them: ```for id in user_ids:```.

create_dendrogram
```

1. You could assign value to `color_threshold` in a single line as follows:

color_treshold = 0.75 if color_treshold is None else color_treshold
```
and similar to the second definition to simplify the nested if - else statements.

2. You could also write the if-else statement in `plt.xticks` in a single line:

 plt.xticks(np.arange(0.0, 1.1, 0.1) if x_max <= 1 else np.arange(0, x_max + 1, 1))
```

_get_cluster_label_for_user
```

1. It seems that `index` is not used within the function, hence `enumerate` and updating its value within the loop (`index += 1`) are not needed.

2. The try-except block adds complexity and it isn't really needed. You could replace it with an if-else block:

if card in df_u["card_label"].values:
    ...
else:
    print(f'"{card}" is not a valid card label. Removed from list.')
    cluster_cards.remove(card)
    print("Continue with cards: %s" % cluster_cards)
```

_get_cards_for_label
```

1. As mentioned above, I would recommend using loc.

get_cluster_labels
```

1. The while loop and the try-except block add complexity and are not necessary here. You could replace the while loop with a for loop after listing all unique user_ids (similar to the suggestion above for `get_distance_matrix`). This way, the try-except block can be removed as we can guarantee that `df_u` is not empty (also getting rid of the `UnboundLocalError` which isn't specific and may be confusing)

2. PEP-8 recommends using format strings instead of concatenation. For example, you could replace the following line:

print("User " + str(id) + " labeled card(s): " + cluster_label)
```
with:

print(f"User {id} labeled card(s): {cluster_label}")
```

get_cluster_labels_df
```

1. Similarly to the previous function, I would recommend using a for loop instead of a while loop and removing the try-except block to simplify the code and improve readability.
