### 2023-05-24 0c3b4dd
Fix _get_distance_matrix_for_user function according to reviewer comments
1) remove redundant updating of index variables, 2) only calculate upper triangle of symmetric matrix and copy second half, 3) use loc instead of query
