+++ src/cardsort/analysis.py
@@ -18,14 +18,9 @@ def _get_distance_matrix_for_user(df_user):
     X = np.zeros((n, n))
     for i in range(n):
         for j in range(n):
-            cat1 = df_user.query("card_id==" + str(i + 1))["category_label"].values[0]
-            cat2 = df_user.query("card_id==" + str(j + 1))["category_label"].values[0]
-            if cat1 == cat2:
-                X[i, j] = 0
-            else:
-                X[i, j] = 1
-            j += i
-        i += 1
+            cat1 = df_user.loc[df_user["card_id"] == (i + 1), "category_label"].values[0]
+            cat2 = df_user.loc[df_user["card_id"] == (j + 1), "category_label"].values[0]
+            X[i, j] = X[j, i] = 0 if cat1 == cat2 else 1
     return X
 
 
