### 2023-05-24 269a52a
Fix get_distance_matrix function according to reviewer comments
replaced while by for loop
