### 2023-05-24 8fc8679
Fix create_dendrogram function according to reviewer comments
simplify nested if-else statements to fit in one line
