### 2023-05-24 7bbdb84
fix _get_cluster_label_for_user according to reviewer comments
1) remove enumerate and index, 2) replace try-except block by if-else, 3) use loc instead of query
