### 2023-05-24 b882486
Fix _get_cards_for_label according to reviewer comments
Use loc instead of query
