### 2023-05-24 e6be892
Fix get_distance_matrix: remove unnecessary manual increment of id
