### 2023-05-24 031ae0d
Fix get_cluster_labels according to reviewer comments
1) Replace while by for loop, 2) Replace try-except by if-else
