### 2023-05-24 2ed2dba
Fix get_cluster_labels_df according to reviewer comments
1) replace while by for loop, 2) replace try-except by if-else
