### 2023-05-24 #102 @katoss
Hi @Robaina, thanks a lot again for the thorough review! 🌟

I have started to address your comments, mainly so far by adapting the functions:
- `_get_distance_matrix_for_user` (commits: [0c3b4dd](https://github.com/katoss/cardsort/commit/0c3b4ddf322cfda5525837ed226a78de6c641235))
- `get_distance_matrix` (commits: [269a52a](https://github.com/katoss/cardsort/commit/269a52ad44ce7f5236f13fbcdf8c4dd5d2397323), [e6be892](https://github.com/katoss/cardsort/commit/e6be8925003ee4d7ec91c4a64201a9a56624f4a9))
- `create_dendrogram` (commits: [8fc8679](https://github.com/katoss/cardsort/commit/8fc8679fdff89e803b8c1b15e220b52a1d6edbfb))
- `_get_cluster_label_for_user` (commits: [7bbdb84](https://github.com/katoss/cardsort/commit/7bbdb8415158e291dac20fc5b771c9389920b46b))
- `_get_cards_for_label` (commits: [b882486](https://github.com/katoss/cardsort/commit/b88248683cbefa26fab0c48752feaa8575e12769))
- `get_cluster_labels` (commits: [031ae0d](https://github.com/katoss/cardsort/commit/031ae0d610187dd3fbb29c4b8ea5d58b92facdca))
- `get_cluster_labels_df` (commits: [2ed2dba](https://github.com/katoss/cardsort/commit/2ed2dbab3d119d3ff9025ba416263fc0785c393e))
- I also added the **link to the docs to the about section** of the repository

Thank you for helping me improve the code quality! Please let me know if I managed to adapt the functions as you imagined.

Regarding the rest of the feedback, I will try to address it as soon as possible.
