### 2023-05-25 #102 @katoss
Ok, thank you! Then my intuition wasn't wrong. I followed [this tutorial](https://py-pkgs.org/) to create the package and CI/CD, and did not question the way they did it, since it was the first time for me setting all of this up.
Is it possible to trigger the whole CD manually so it only publishes a new version when I want it to? Or maybe the easiest thing would be that I create a development branch and collect the changes there, as you said.
