### 2023-05-25 #102 @Robaina
Perhaps the easiest solution is to trigger the PIP upload only on release --- automatically triggered after a new release of the repo (with a new version number). I could do a PR with this if you like.
