### 2023-05-25 #102 @Robaina
Interesting, I was thinking of using:

```
on:
  release:
    types: [published]
```

This should trigger the workflow only after a new release is created. Either way should work, though!
