### 2023-05-25 #102 @Batalex
Both approaches should work. The "release published" one would work on all kinds of releases, whereas the tags one is limited by the regexp. The latter has the benefit of being actionable with basic git commands:
```
git tag x.y.z
git push origin x.y.z
```

In both cases, we would need to run semantic-release ourselves to get the proper "x.y.z" to use
