### 2023-05-28 #102 @Robaina
Hi @katoss ,

great work! I reviewed your commits and left some minor comments. Only one related to efficiency in _get_distance_matrix_for_user, you forgot to iterate only over the upper triangle.

I suggest that you use GitHub issues and pull requests in future interactions, it's easier to get code reviewed by a peer this way. Also, the code can be reviewed before a merge is done (for instance from a dev branch to main).

Cheers
