### 2023-05-30 #102 @katoss
>

Thank you @Robaina ! I fixed the function in a new branch. I invited you and @Batalex as collaborators so I can assign you as reviewers to a pull request
