### 2023-05-30 #102 @katoss
And with regards to the cd, the line that will have to be changed is this one, and the rest can stay the same?

` # Only run this job if new work is pushed to the "main" branch`
` if: github.event_name == 'push' && github.ref == 'refs/heads/main'`
