### 2023-05-30 #102 @Batalex
> And with regards to the cd, the line that will have to be changed is this one, and the rest can stay the same?

With this configuration, you would make a new release each time `main` is updated. You might want to restrict releases to git tags or github releases (based on git tags)
