### 2023-05-31 #102 @khynder
Hi @katoss @Batalex @Robaina !

Here is my review:
First of all, I really like the idea behind this package, the use of a dendrogram to visualize the card sorting is inteligent and usefull. The package is well documented and easy to read and use! Congrats 🎉.

- [ ] **Vignette(s)** demonstrating major functionality that runs successfully locally. _@Batalex what are the "vignettes" ?_

- [ ] The package name _the title of the README is not exactly the package name_
- [ ] Badges for: _no badges for now_

- [x] Package documentation is clear and easy to find and use. 🎉🎉🎉

- [x] **Installation:** Installation succeeds as documented. _I installed directly the last version_
- [ ] **Functionality:** Any functional claims of the software been confirmed. _Be careful with the data format when it does not come directly from [kardsort.com](https://kardsort.com/), more details below_
  - [ ] Tests cover essential functions of the package and a reasonable range of inputs and conditions. _Maybe a test could be added on the order of labels in the dendrogram (on the test-data.csv)_
    - [ ] Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass) _no linter_

---

### General remarks:

It would be safer if a function was added to check that the data format is correct, especially if the package is used with data which does not come from kardsort.com. More specifically:
-	The first user in df should have 1 as user_id (otherwise in get_distance_matrix, distance_matrix_all will be used before creation, and labels could not be computed)
-	All users should categorize all cards and each card only once
-	card_id should go from 1 to n (n being the number of cards)
-	one card_id should always be associated with exactly one card_label (not two or more)

### Function by function:

#### get_distance_matrix_for_user:
It can be written as a comprehension list if you want (no obligation, if you feel more at ease with the current code keep it 😊)
df_user = df_user.set_index("card_id")
X = np.array(
    [
        [
            df_user.loc[i + 1, "category_label"]
            != df_user.loc[j + 1, "category_label"]
            for i in range(n)
        ]
        for j in range(n)
    ]
).astype(int)
```

#### get_distance_matrix:
- avoid using `print` in code, if you have a lot of users in the database it could slow down the execution. You can add a Boolean parameter `print_progression` or use loggers (@Batalex could explain loggers better than me I think)
- the first user should have `user_id = 1`, otherwise `distance_matrix_all` won’t exist. It can stay that way, with a check at the beginning of the function to stop the computation if the first user_id is not 1. Or you can create before the for loop a variable `distance_matrix_all = None,` and replace the `id == 1` condition by a `distance_matrix_all is not None` condition

#### create_dendrogram:
when you define labels
labels = df.loc[df["user_id"] == 1]["card_label"].squeeze().to_list()
```
It is very important that the order of the card_label in labels is exactly the same as the one used to create the distance_matrix. The distance_matrix is created with the card_id in the order 1, 2, 3, 4, …, n. So labels should be: first the card_label corresponding to card_id = 1, then card_label corresponding to card_id = 2, then card_label corresponding to card_id = 3, etc. The above code line works fine if the card_id for the user_id 1 are in the order 1, 2, 3 etc, but does not work if the card_id are shuffled.
I’m not sure if i'm clear, so I tried with an example: if you replace the user_id results in test_data by:

card_id | card_label | category_id | category_label | user_id
-- | -- | -- | -- | --
9 | Mooncake | 4 | Moon-shaped | 1
10 | Moon | 4 | Moon-shaped | 1
1 | Dog | 1 | pets | 1
2 | Tiger | 1 | pets | 1
3 | Cat | 1 | pets | 1
4 | Apple | 2 | lunch | 1
5 | Sandwich | 2 | lunch | 1
6 | Banana | 3 | long food | 1
7 | Hot Dog | 3 | long food | 1
8 | Croissant | 4 | Moon-shaped | 1

Then the dendrogram will be:
which makes no sense.

One way to correct this behavior is to add a `sort_values(“card_id”)` to the labels definition.
labels = df.loc[df["user_id"] == 1].sort_values(“card_id”)["card_label"].squeeze().to_list()
```
Note that it should also be changed in the `example.ipynb`.

#### get_cluster_labels:
-	I think the if/else condition after the for loop can be removed.
-	There is a lot of common code lines between `get_cluster_labels` and `get_cluster_labels_df`. These two functions can be factorized into one function, with 2 boolean parameters to indicate what type of return you want, for example `print_results` and `return_df_results`.

It could look like this:
def get_cluster_labels(df, cluster_cards, print_results=True, return_df_results=False):
    if return_df_results:
        cluster_df = pd.DataFrame(columns=["user_id", "cluster_label", "cards"])

    user_ids = df["user_id"].unique()

    for id in user_ids:
        df_u = df.loc[df["user_id"] == id]
        cluster_label = _get_cluster_label_for_user(df_u, cluster_cards)
        if cluster_label is not None:
            if print_results:
                print("User " + str(id) + " labeled card(s): " + cluster_label)
            if return_df_results:
                cards = _get_cards_for_label(cluster_label, df_u)
                cluster_df = pd.concat(
                    [
                        cluster_df,
                        pd.DataFrame.from_records(
                            [
                                {
                                    "user_id": id,
                                    "cluster_label": cluster_label,
                                    "cards": cards,
                                }
                            ]
                        ),
                    ],
                    ignore_index=True,
                )
        else:
            if print_results:
                print("User " + str(id) + " did not cluster cards together.")

    if return_df_results:
        return cluster_df
```

- the warning "is/are not a valid card label" can be moved in this function, instead of being checked at a `user_id` level:

if not set(cluster_cards) <= set(df["card_label"]):
    missing_card_labels = set(cluster_cards) - set(df["card_label"])
    print(
        f'"{missing_card_labels}" is/are not a valid card label. Removed from list.'
    )
    cluster_cards = [
        card_label
        for card_label in cluster_cards
        if card_label not in missing_card_labels
    ]
    print("Continue with cards: %s" % cluster_cards)
```

#### get_cluster_label_for_user

It can be simplified with the “isin()” pandas method:
https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.isin.html

    list_cat = df_u.loc[
        df_u["card_label"].isin(cluster_cards), "category_label"
    ].unique()
    if len(list_cat) == 1:
        return list_cat.squeeze()
```

A “)” is missing at the end of the line starting with "distance_matrix", in the “Adapting the dendrogram” section.
