### 2023-06-01 #102 @Batalex
> If I understood right, this is done retrospectively, e.g. I decide at some point that everything up to commit abcdef123456 shall be included in the new release

Yes, that would be the process
