### 2023-06-01 #102 @Batalex
Thanks @khynder, I knew you would like this package 🤗

@katoss, I have a few remarks as well. This is not a thorough review, but I like to give my opinion even as the editor.

- imports should be at the top level, `__all__` should go below. I would advice as suggested by @Robaina using a code formater and a linter
- `_get_distance_matrix_for_user` could be rewritten using scipy's `pdist`, but it is worth checking if the rewrite to compute the upper triangular matrix is more efficient.
- `get_cluster_labels`, `get_distance_matrix` and `get_cluster_labels_df` use the form `id = 1; while id <= max(df.user.id)`. We can make this more robust by using a for loop on the values in the dataframe. The original version would fail for string ids or even if a value is missing, e.g `[1, 2, 4, 5]`. Since the value `1` is used to initialize the distance matrix, you may add a boolean condition (to be changed in the first iteration), or you may create a matrix with `np.zeros` before the for loop. The author may use the form `for user_id, sub_df in df.groupby("user_id"):`.
- Since the project targets python versions >= 3.8, I would advise using f-strings to interpolate variables in strings.
- Most of the time, we want to have the less possible code in try .. except blocks. The more specific you are when catching exception, the easier it is to debug your program.
- I believe most of the complexity of `get_cluster_labels` could be cut down by checking the values of `cluster_cards` against the possible values before entering the loop. It is quite risky to mutate a datastructure (especially removing values) while using it in a loop, and usually considered a bad practice.
