+++ .github/workflows/ci-cd.yml
@@ -38,8 +38,11 @@ jobs:
     # Only run this job if the "ci" job passes
     needs: ci
 
-    # Only run this job if new work is pushed to the "main" branch
-    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+    # Only run this job when push event involves release tag
+    on:
+      push:
+        tags:
+        - 'v[0-9]+.[0-9]+.[0-9]+'
 
     # Set up operating system
     runs-on: ubuntu-latest
