### 2023-06-01 #102 @katoss
> Ok @Batalex, thank you for the clarification. I have never worked with tags, but am trying to understand them.
>
> Is it right that I would need to put this code in the CD pipeline:
>
> ```
> on:
>   push:
>     tags:
>       - 'v[0-9]+.[0-9]+.[0-9]+'
> ```
>
> instead of
>
> ```
>  # Only run this job if new work is pushed to the "main" branch
>  if: github.event_name == 'push' && github.ref == 'refs/heads/main'
> ```
>
> ?
>
> And then I can assign tags to commits in the command line? If I understood right, this is done retrospectively, e.g. I decide at some point that everything up to commit `abcdef123456` shall be included in the new release, so I'd do `git tag v1.0.0 abcdef123456`, push the tag, and then the release should be triggered?

Just created a [PR with these changes](https://github.com/katoss/cardsort/pull/1) :)
