+++ .github/workflows/ci-cd.yml
@@ -42,7 +42,7 @@ jobs:
     on:
       push:
         tags:
-        - 'v[0-9]+.[0-9]+.[0-9]+'
+          - 'v[0-9]+.[0-9]+.[0-9]+'
 
     # Set up operating system
     runs-on: ubuntu-latest
