### 2023-06-02 a33fa6a
fix cd by replacing trigger keyword 'on' with 'if' condition
