+++ .github/workflows/ci-cd.yml
@@ -39,10 +39,7 @@ jobs:
     needs: ci
 
     # Only run this job when push event involves release tag
-    on:
-      push:
-        tags:
-          - 'v[0-9]+.[0-9]+.[0-9]+'
+    if: startsWith(github.ref, 'refs/tags/v')
 
     # Set up operating system
     runs-on: ubuntu-latest
