### 2023-06-02 #102 @Robaina
Hi @katoss, @khynder, @Batalex,

I checked the [PR](https://github.com/katoss/cardsort/pull/1) above. I left some comments. I'm not entirely sure that the current solution will work as intended. Would appreciate it if you take a look.
