### 2023-06-02 #102 @katoss
Just commented. And I second @Robaina, would be great if you could have a look over our solution ideas, @Batalex and @khynder !
