### 2023-06-14 36429e2
Fix: separate CI and CD to trigger CD only on release
