### 2023-06-14 #102 @katoss
Just pushed a commit and commented on the [PR](https://github.com/katoss/cardsort/pull/1) :)
Will have a look at the other feedback now.
