+++ pyproject.toml
@@ -12,6 +12,7 @@ pandas = "^1.5.0"
 numpy = "^1.20.3"
 matplotlib = "^3.3.2"
 scipy = "^1.7.1"
+black = "^23.3.0"
 
 [tool.poetry.dev-dependencies]
 
