### 2023-06-14 0fe257e
Add type hints
