### 2023-06-14 #102 @katoss
Made another [draft PR](https://github.com/katoss/cardsort/pull/2) for some **general formatting issues** (linting, type hints, etc).

@Batalex and @khynder, could you accept my invite to be collaborator in this repo? Then I can add you as reviewers to PRs.

So as an overview, as far as I see the **main blocks of feedback to address** are the CI/CD improvements, general formatting, improvement of functions, and documentation. I think I will wait with further function improvements until CI/CD and general formatting blocks (the two open PRs) are fixed, but will be able to address documentation quite independently.

Thanks again for your reviews, great to see the package progessing! :)
