### 2023-06-20 85b8873
Fix CD: change to 'release type [published]' and remove redundant conditions
