+++ .github/workflows/CD.yml
@@ -7,17 +7,10 @@ on:
       - completed
   
   release:
-  types:
-    - created
+    types: [published]
 
 jobs:
   cd:
-    # Only run this job if the "ci" job passes
-    needs: ci
-
-    # Only run this job when push event involves release tag
-    if: startsWith(github.ref, 'refs/tags/v')
-
     # Set up operating system
     runs-on: ubuntu-latest
 
