### 2023-06-20 233f220
Move documentation building from CI to CD
