+++ .github/workflows/CI.yml
@@ -30,6 +30,3 @@ jobs:
       uses: codecov/codecov-action@v3
       with:
         files: ./coverage.xml # coverage report
-
-    - name: Build documentation
-      run: poetry run make html --directory docs/
