### 2023-06-20 5defea8
Fix dependencies: move black to dev dependencies
