+++ pyproject.toml
@@ -12,7 +12,6 @@ pandas = "^1.5.0"
 numpy = "^1.20.3"
 matplotlib = "^3.3.2"
 scipy = "^1.7.1"
-black = "^23.3.0"
 
 [tool.poetry.dev-dependencies]
 
@@ -24,6 +23,7 @@ sphinx-autoapi = "^2.0.1"
 sphinx-rtd-theme = "^1.2.0"
 python-semantic-release = "^7.33.2"
 myst-nb = "^0.17.2"
+black = "^23.3.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
