### 2023-06-20 1024772
Fix _get_distance_matrix_for_user: Only iterate over upper triangle
