### 2023-06-21 0a2b4b1
Add and configure pre-commit hook for black
