+++ pyproject.toml
@@ -24,6 +24,7 @@ sphinx-rtd-theme = "^1.2.0"
 python-semantic-release = "^7.33.2"
 myst-nb = "^0.17.2"
 black = "^23.3.0"
+pre-commit = "^3.3.3"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
