### 2023-06-21 93b9bf1
Add link to contributing guidelines doc
