### 2023-06-21 d2b3642
Add logo to readme and extend package description
