+++ README.md
@@ -1,6 +1,12 @@
-# cardsort
+<br>
+<div align="center">
+<img src="https://github.com/katoss/cardsort/blob/documentation/logo.png" width="400">
+</div>
+<hr>
 
-A package that helps UX researchers quickly analyse data from card sorting exercises. Click [here](https://www.nngroup.com/articles/card-sorting-definition/) to learn more about the card sorting method.
+# A Python package to cluster and visualize data from open card sorting tasks
+
+Cardsort helps UX researchers quickly analyse data from card sorting exercises using hierarchical cluster analysis. This task helps to understand how people organize information, and is frequently used to develop information architectures for websites. Click [here](https://www.nngroup.com/articles/card-sorting-definition/) to learn more about the card sorting method.
 
 __More precisely, it helps you to:__
 * Create dendrograms
