### 2023-06-21 b791460
Adapt list of functionality, to make it more precise
