### 2023-06-21 1545435
Add path to example data to readme
