+++ README.md
@@ -47,6 +47,9 @@ import pandas as pd
 path = "example-data.csv" # data with columns: card_id, card_label, category_id, category_label, user_id
 df = pd.read_csv(path) 
 ```
+
+The data used in this example can be found under [/docs/example-data.csv](https://github.com/katoss/cardsort/blob/main/docs/example-data.csv)
+
 __Create a dendrogram that summarizes user-generated clusters__
 ```python
 analysis.create_dendrogram(df)
