### 2023-06-23 ec3ae25
Add badges for pypi, ci, codecov, python version
