### 2023-06-23 #102 @katoss
Hey, I just made a [PR](https://github.com/katoss/cardsort/pull/3) for improvements of the documentation @Robaina @khynder @Batalex  :)

And I converted the  [linting/code formatting PR](https://github.com/katoss/cardsort/pull/2) from draft to PR, I think it should be ready to be merged.
