+++ tests/test_analysis.py
@@ -6,53 +6,12 @@ df = pd.read_csv("tests/test-data.csv")
 
 
 def test_get_distance_matrix():
-    expected = [
-        2.0,
-        1.0,
-        5.0,
-        5.0,
-        5.0,
-        4.0,
-        5.0,
-        5.0,
-        5.0,
-        1.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        4.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        5.0,
-        2.0,
-        1.0,
-        4.0,
-        3.0,
-        3.0,
-        5.0,
-        3.0,
-        2.0,
-        1.0,
-        1.0,
-        5.0,
-        3.0,
-        3.0,
-        3.0,
-        5.0,
-        2.0,
-        2.0,
-        5.0,
-        0.0,
-        4.0,
-        4.0,
-    ]
+    # fmt: off
+    expected = [2., 1., 5., 5., 5., 4., 5., 5., 5., 1., 5., 5., 
+                5., 5., 5., 5., 4., 5., 5., 5., 5., 5., 5., 5.,
+                2., 1., 4., 3., 3., 5., 3., 2., 1., 1., 5., 3., 
+                3., 3., 5., 2., 2., 5., 0., 4., 4.]
+    # fmt: on
     actual = analysis.get_distance_matrix(df)
     assert (actual == expected).all(), "Distance matrix not correct"
 
