### 2023-07-05 934f6f7
Change logging level to info
