### 2023-07-05 d43f500
Check if first user id is 1 in get_distance_matrix
