### 2023-07-05 #102 @katoss
Hi,
I just started a [draft PR for changes in functions](https://github.com/katoss/cardsort/pull/5) based on the reviews of @khynder and @Batalex :)

It is quite a lot, I listed everything, but have addressed only some issues so far, and have some questions for others. @khynder , could you accept the invitation to become a collaborator on cardsort, so I can assign you as a reviewer to the PR?

I also feel there might be still an issue with the trigger for CD, since CD started for all commits for patch changes in the new PR (see [Actions](https://github.com/katoss/cardsort/actions)).
