+++ .github/workflows/CD.yml
@@ -5,6 +5,7 @@ on:
     workflows: [CI]
     types:
       - completed
+    branches: [main]
   
   release:
     types: [published]
