### 2023-07-12 3f1b18e
Fix logging, add nullHandler to init
