### 2023-07-12 76ed2da
Simplify _get_cluster_label_for_user with isin(), and remove test for correct card labels from this function
