### 2023-07-12 ab1d72c
Merge get_cluster_labels and get_cluster_labels_df into one function
