### 2023-07-12 caa2cb7
Remove get_cluster_labels_df function
