+++ src/cardsort/analysis.py
@@ -270,86 +270,3 @@ def get_cluster_labels(
 
     if return_df_results:
         return cluster_df
-
-
-"""
-    cluster_labels = []
-    user_ids = df["user_id"].unique()
-
-    for id in user_ids:
-        df_u = df.loc[df["user_id"] == id]
-        if len(cluster_cards) > 0:
-            cluster_label = _get_cluster_label_for_user(df_u, cluster_cards)
-            if cluster_label is not None:
-                logging.info(f"User {id} labeled card(s): {cluster_label}")
-                cluster_labels.append(cluster_label)
-            else:
-                logging.info(f"User {id} did not cluster cards together.")
-        else:
-            logging.info("No cards left in list.")
-            break
-    return cluster_labels
-"""
-
-
-def get_cluster_labels_df(df: pd.DataFrame, cluster_cards: List[str]) -> pd.DataFrame:
-    """
-    Return category labels and user id for each user who clustered given list of cards together.
-    Also returns full list of cards in that category.
-
-    Parameters
-    ----------
-    df : pandas.DataFrame
-        Columns:
-                Name: card_id, dtype: int64
-                Name: card_label, dtype: object
-                Name: category_id, dtype: int64
-                Name: category_label, dtype: object
-                Name: user_id, dtype: int64
-        These columns correspond to the 'Casolysis Data (.csv) - Recommended' export from kardsort.com.
-
-    cluster_cards : list of str
-        List of card-labels for which you would like to get user-generated cluster-labels.
-
-    Returns
-    -------
-    out : pandas.DataFrame
-        Columns:
-            Name: user_id, int
-            Name: cluster_label, str
-            Name: cards, list of str
-        Dataframe with one row for each user who clustered the given cards together, including category label and
-        the full list of cards in that category.
-    """
-
-    cluster_df = pd.DataFrame(columns=["user_id", "cluster_label", "cards"])
-    user_ids = df["user_id"].unique()
-
-    for id in user_ids:
-        df_u = df.loc[df["user_id"] == id]
-        if len(cluster_cards) > 0:
-            cluster_label = _get_cluster_label_for_user(df_u, cluster_cards)
-            if cluster_label is not None:
-                logging.info(f"User {id} labeled card(s): {cluster_label}")
-                cards = _get_cards_for_label(cluster_label, df_u)
-                cluster_df = pd.concat(
-                    [
-                        cluster_df,
-                        pd.DataFrame.from_records(
-                            [
-                                {
-                                    "user_id": id,
-                                    "cluster_label": cluster_label,
-                                    "cards": cards,
-                                }
-                            ]
-                        ),
-                    ],
-                    ignore_index=True,
-                )
-            else:
-                logging.info(f"User {id} did not cluster cards together.")
-        else:
-            logging.info("No cards left in list.")
-            break
-    return cluster_df
