### 2023-07-12 a9d9ef9
Add check if there are still cards in list. Otherwise return None
