### 2023-07-12 0b2a65f
Adapt tests to adapted version of get_cluster_labels
