### 2023-07-19 3867cfc
Change logging to logger, change remaining prints to logger, change default return_df to true
