### 2023-07-19 1e02db6
Fix test_get_cluster_labels_print by adding return_df_results=False
