### 2023-07-19 a3e30be
Created _check_data function to check input format and added it to each function that accepts dfs from users
