+++ src/cardsort/analysis.py
@@ -17,6 +17,40 @@ __all__ = [
 logger = logging.getLogger(__name__)
 
 
+def _check_data(df: pd.DataFrame) -> bool:
+    # check if user_ids go from 1 to n=number of users
+    user_ids = df["user_id"].unique()
+    for i, user_id in enumerate(user_ids):
+        if user_id != i + 1:
+            logger.error("User_ids are not in the right order or contain gaps.")
+            return False
+    # check if each card_id is always associated with exactly one card_label
+    card_id_counts = df.groupby("card_id")["card_label"].nunique()
+    for card_id, count in card_id_counts.items():
+        if count != 1:
+            logger.error(
+                f"Card_id {card_id} is associated with {count} different card_labels."
+            )
+            return False
+    # check if card_ids go from 1 to n=number of cards
+    card_ids = df["card_id"].unique()
+    for i, card_id in enumerate(card_ids):
+        if card_id != i + 1:
+            logger.error("Card_ids are not in the right order or contain gaps.")
+            return False
+    # check if all users categorize each card exactly once
+    categorized_cards_per_user = (
+        df.groupby("user_id")["card_id"].count().reset_index(name="card_count")
+    )
+    card_count_equal_categorized = (
+        categorized_cards_per_user["card_count"] == len(card_ids)
+    ).all()
+    if not card_count_equal_categorized:
+        logger.error("Not all users have categorized each card exactly once.")
+        return False
+    return True
+
+
 def _get_distance_matrix_for_user(df_user: pd.DataFrame) -> np.ndarray:
     n = len(df_user)
 
@@ -54,9 +88,13 @@ def get_distance_matrix(df: pd.DataFrame) -> np.ndarray:
         A condensed distance matrix (a flat array containing the upper triangle of a distance matrix)
         representing the pairwise similarity of all cards.
     """
-    user_ids = df["user_id"].unique()
-
-    if user_ids[0] == 1:
+    if _check_data(df) == False:
+        logger.error(
+            "The DataFrame does not correspond to the required format. No distance matrix generated."
+        )
+        return None
+    else:
+        user_ids = df["user_id"].unique()
         for id in user_ids:
             df_u = df.loc[df["user_id"] == id]
             logger.info(f"Computing distance matrix for user {id}")
@@ -67,9 +105,6 @@ def get_distance_matrix(df: pd.DataFrame) -> np.ndarray:
                 distance_matrix_all = np.add(distance_matrix_all, distance_matrix_user)
         condensed_distance_matrix = squareform(distance_matrix_all)
         return condensed_distance_matrix
-    else:
-        logger.ERROR("The first user ID needs to equal 1")
-        return None
 
 
 def create_dendrogram(
@@ -121,46 +156,55 @@ def create_dendrogram(
         Can be a fraction (0 - 1) or an absolute value (<= n = number of users).
         The default cut is at 75%.
     """
-
-    if distance_matrix is None:
-        distance_matrix = get_distance_matrix(df)
-
-    count_types = ["absolute", "fraction"]
-    if count not in count_types:
-        raise ValueError("Invalid count type. Expected one of: %s" % count_types)
-
-    linkage_types = ["average", "complete", "single"]
-    if linkage not in linkage_types:
-        raise ValueError("Invalid linkage. Expected one of: %s" % linkage_types)
-
-    if count == "fraction":
-        distance_matrix = distance_matrix / np.max(distance_matrix)
-        color_threshold = 0.75 if color_threshold is None else color_threshold
+    if _check_data(df) == False:
+        logger.error(
+            "The DataFrame does not correspond to the required format. No dendrogram generated."
+        )
+        return None
     else:
-        color_threshold = (
[…]
