### 2023-07-19 926a0bf
Add test for _check_data function
