### 2023-07-19 4bffebe
Adapt usage example to changes in get_cluster_labels function
