### 2023-07-19 #102 @katoss
I worked on the `fix-functions` [PR](https://github.com/katoss/cardsort/pull/5), I think it might be good to go @khynder @Batalex
