### 2023-07-20 9c629a9
Replace for loop with faster numpy array comparison in _get_distance_matrix_for_user
