### 2023-07-27 192bda9
Replace 'id' variable name with 'id_' to prevent potential issues with built-in 'id'
