### 2023-07-27 0cf44a0
Append to list to concat list only once after loop
