### 2023-07-27 ff42545
Adapt test to changes in get_cluster_labels
