### 2023-07-27 c9f27d7
Remove check if card_ids go from 1 to n in _check_data
