### 2023-08-01 323d094
Move License, Contributing, and Credits under Project information heading
