### 2023-08-01 5e5a5b9
Add contributors list config code and badge
