### 2023-08-01 #102 @katoss
I merged the `fix-functions` PR 🚀 , thank you for your help @khynder @Batalex .
I also added the repo to zenodo and added a badge, as you suggested @Robaina (on branch `update-readme` so far). I'm also about to add a list of contributors, will try to figure out how to use the all-contributors bot in the next days. I should probably also change the repo-status from WIP to active?

I think I worked through all the feedback now 🤔 What are the next steps? :)
