### 2023-08-04 61800e8
Move automatically generated all-contributors badge to the right section
