+++ README.md
@@ -123,6 +123,7 @@ Thanks goes to these wonderful people ([emoji key](https://allcontributors.org/d
       <td align="center" valign="top" width="14.28%"><a href="https://tzovar.as"><img src="https://avatars.githubusercontent.com/u/674899?v=4?s=100" width="100px;" alt="Bastian Greshake Tzovaras"/><br /><sub><b>Bastian Greshake Tzovaras</b></sub></a><br /><a href="#fundingFinding-gedankenstuecke" title="Funding Finding">🔍</a> <a href="https://github.com/katoss/cardsort/commits?author=gedankenstuecke" title="Code">💻</a> <a href="#ideas-gedankenstuecke" title="Ideas, Planning, & Feedback">🤔</a> <a href="#mentoring-gedankenstuecke" title="Mentoring">🧑‍🏫</a></td>
       <td align="center" valign="top" width="14.28%"><a href="http://batalex.github.io"><img src="https://avatars.githubusercontent.com/u/11004857?v=4?s=100" width="100px;" alt="Alex Batisse"/><br /><sub><b>Alex Batisse</b></sub></a><br /><a href="https://github.com/katoss/cardsort/pulls?q=is%3Apr+reviewed-by%3ABatalex" title="Reviewed Pull Requests">👀</a> <a href="#ideas-Batalex" title="Ideas, Planning, & Feedback">🤔</a></td>
       <td align="center" valign="top" width="14.28%"><a href="https://www.linkedin.com/in/semidan-robaina/"><img src="https://avatars.githubusercontent.com/u/21340147?v=4?s=100" width="100px;" alt="Semidán Robaina"/><br /><sub><b>Semidán Robaina</b></sub></a><br /><a href="https://github.com/katoss/cardsort/pulls?q=is%3Apr+reviewed-by%3ARobaina" title="Reviewed Pull Requests">👀</a> <a href="#ideas-Robaina" title="Ideas, Planning, & Feedback">🤔</a></td>
+      <td align="center" valign="top" width="14.28%"><a href="https://github.com/khynder"><img src="https://avatars.githubusercontent.com/u/133049827?v=4?s=100" width="100px;" alt="khynder"/><br /><sub><b>khynder</b></sub></a><br /><a href="https://github.com/katoss/cardsort/pulls?q=is%3Apr+reviewed-by%3Akhynder" title="Reviewed Pull Requests">👀</a> <a href="#ideas-khynder" title="Ideas, Planning, & Feedback">🤔</a></td>
     </tr>
   </tbody>
 </table>
