### 2023-08-16 #102 @Batalex
🎉 `cardsort` has been approved by pyOpenSci! Thank you @katoss  for submitting `cardsort` and many thanks to @Robaina and @khynder  for reviewing this package! 🐈‍⬛

## Author Wrap Up Tasks

- [ ] Add the badge for pyOpenSci peer-review to the README.md of `cardsort`. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/102)`.

@Robaina, @khynder we would greatly appreciate if you could fill the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7) as well!
If any of you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and documentation in the [peer-review-guide](https://www.pyopensci.org/software-peer-review).
