### 2023-08-22 8327603
Separate definition of all-contributors badge from logo display to avoid linebreak
