### 2023-08-22 #102 @katoss
Amazing! I just added the badge and did the survey. Thanks a lot for spending the time on reviewing everything so thoroughly @Batalex @Robaina and @khynder! The review was really helpful and instructive, and improved the package a lot :)
