### 2023-08-23 #102 @Batalex
Congrats @katoss, I am glad this review went so well. I am truly pleased that we were able to add value to your work.
Best of luck with your thesis! 🤞
I will be closing this submission issue now.

We offer authors to join us on our slack / [discourse](https://pyopensci.discourse.group/). Let me know if you are interested in joining the slack.
We can also promote `cardsort` with a blog post, if you feel like you have not written enough with your thesis 😁 [This](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation) is what it could look like.
