### 2023-09-01 #102 @katoss
Amazing, thanks @Batalex :) And yes, I would like to join the slack. Could you send me an invite?
A blog article sounds like a good idea, I will try to write a draft some time this month.

I saw that the [PR](https://github.com/pyOpenSci/pyopensci.github.io/pull/267) to update the pyOpenSci website with the `cardsort` package is still open and has no reviewer assigned. Is normal that it just takes a while, or do you think it would be better if I send a reminder?
