### 2023-09-03 #102 @Batalex
I just invited you 🐈‍⬛

> Is normal that it just takes a while, or do you think it would be better if I send a reminder?

It seems that we have an issue with pre-commit.ci, but it should be fixed soon
