### 2023-09-11 #102 @lwasser
hey there @katoss 👋 thank you for mentioning this issue in our website update pr!! i saw it. it took me a while to merge as there were some bugs in the build that i wanted to fix before merging. it would have broken the website! but generally these will get merged more quickly - as that build becomes a bit more stable.

cardsort is [now listed on our website](https://www.pyopensci.org/python-packages.html)!! 🚀 and we look forward to seeing you around slack!!

thank you everyone for your work on this pr!! and @Batalex for leading the charge on the review ✨  !!!
